backend/requirements.txt
flask
flask-cors
ultralytics
torch
Pillow
google-generativeai 